// Zybach.Web/src/app/shared/generated/model/report-template-dto.ts
/**
 * Zybach
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReportTemplateModelDto } from '././report-template-model-dto';
import { ReportTemplateModelTypeDto } from '././report-template-model-type-dto';
import { FileResourceDto } from '././file-resource-dto';

export class ReportTemplateDto { 
    ReportTemplateID?: number;
    FileResource?: FileResourceDto;
    DisplayName?: string;
    Description?: string;
    ReportTemplateModelType?: ReportTemplateModelTypeDto;
    ReportTemplateModel?: ReportTemplateModelDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}