// Zybach.Web/src/app/shared/generated/model/well-inspection-summary-dto.ts
/**
 * Zybach
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { WellSimpleDto } from '././well-simple-dto';

export class WellInspectionSummaryDto { 
    Well?: WellSimpleDto;
    HasWaterLevelInspections?: boolean;
    LatestWaterLevelInspectionDate?: string;
    HasWaterQualityInspections?: boolean;
    LatestWaterQualityInspectionDate?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}